move app lock to file
