+++ wire-ios-sync-engine/Tests/Source/SessionManager/SessionManagerTests.swift
@@ -457,75 +457,6 @@ extension IntegrationTest {
 
 // MARK: - Account Deletion
 
-final class SessionManagerTests_AppLock: IntegrationTest {
-
-    private var appLock: MockAppLock!
-
-    override func setUp() {
-        super.setUp()
-        // Mock transport doesn't support multiple accounts at the moment so we pretend to be offline
-        // in order to avoid the user session's getting stuck in a request loop.
-        mockTransportSession.doNotRespondToRequests = true
-        appLock = MockAppLock()
-    }
-
-    override func tearDown() {
-        sessionManager!.tearDownAllBackgroundSessions()
-        appLock = nil
-        mockTransportSession.doNotRespondToRequests = false
-        super.tearDown()
-    }
-
-    func test_ItBeginsAppLockTimer_WhenChangingTheActiveUserSession() {
-        // Given
-        let account1 = addAccount(name: "Account 1", userIdentifier: currentUserIdentifier)
-        let account2 = addAccount(name: "Account 2", userIdentifier: .create())
-
-        weak var session1: ZMUserSession?
-        sessionManager?.loadSession(for: account1, completion: { (session) in
-            session1 = session
-            session1?.appLockController = self.appLock
-        })
-
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-        XCTAssertEqual(sessionManager!.activeUserSession, session1)
-
-        // When
-        let switchedAccount = customExpectation(description: "switched account")
-        sessionManager?.select(account2, completion: { _ in
-            switchedAccount.fulfill()
-        }, tearDownCompletion: nil)
-
-        XCTAssertTrue(waitForCustomExpectations(withTimeout: 0.5))
-
-        // Then
-        XCTAssertEqual(appLock.methodCalls.beginTimer.count, 1)
-    }
-
-    func test_ItBeginAppLockTimer_WhenTheAppResignsActive() {
-        // Given
-        let account1 = addAccount(name: "Account 1", userIdentifier: currentUserIdentifier)
-
-        // Make account 1 the active session.
-        weak var session1: ZMUserSession?
-        sessionManager?.loadSession(for: account1, completion: { (session) in
-            session1 = session
-            session1?.appLockController = self.appLock
-        })
-
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-        XCTAssertEqual(sessionManager!.activeUserSession, session1)
-
-        // When
-        let notification = Notification(name: UIApplication.willResignActiveNotification)
-        sessionManager!.applicationWillResignActive(notification)
-
-        // Then
-        XCTAssertEqual(appLock.methodCalls.beginTimer.count, 1)
-    }
-
-}
-
 extension NSManagedObjectContext {
     func createSelfUserAndSelfConversation() {
         let selfUser = ZMUser.selfUser(in: self)
